# server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealForge API",
    "description": "AI-powered meeting intelligence and sales CRM API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API server" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Read-only API key (generated in Settings → API Access). Available on Pro."
      }
    },
    "schemas": {
...
      "EmailDraft": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "draft": {
            "type": "object",
            "properties": {
              "subject": { "type": "string" },
              "body": { "type": "string" }
            }
          }
        }
      },
...
      "EmailSendResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "from": { "type": "string" },
          "to": { "type": "string" }
        }
      },
      "ApiKeyNew": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "key": { "type": "string", "description": "Full API key. Shown only once — store it securely." },
          "keyHash": { "type": "string" },
          "name": { "type": "string" },
          "prefix": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ApiKeySummary": {
        "type": "object",
        "properties": {
          "keyHash": { "type": "string" },
          "name": { "type": "string" },
          "prefix": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true },
          "revoked": { "type": "boolean" }
        }
      },
      "SyncPayload": {
        "type": "object",
        "properties": {
          "meetings": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": { "type": "string" },
                "title": { "type": "string" },
                "startTime": { "type": "string" },
                "endTime": { "type": "string" },
                "duration": { "type": "number" },
                "status": { "type": "string" }
              }
            }
          },
          "analyses": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "meetingId"],
              "properties": {
                "id": { "type": "string" },
                "meetingId": { "type": "string" },
                "summary": { "type": "string" },
                "actionItems": { "type": "array", "items": { "type": "string" } },
                "leadScore": { "type": "number" },
                "modelUsed": { "type": "string" },
                "analyzedAt": { "type": "string" }
              }
            }
          },
          "leads": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": { "type": "string" },
                "meetingId": { "type": "string" },
                "name": { "type": "string" },
                "email": { "type": "string" },
                "company": { "type": "string" },
                "role": { "type": "string" },
                "score": { "type": "number" },
                "stage": { "type": "string" },
                "createdAt": { "type": "string" },
                "updatedAt": { "type": "string" }
              }
            }
          },
          "deals": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": { "type": "string" },
                "leadId": { "type": "string" },
                "title": { "type": "string" },
                "stage": { "type": "string" },
                "value": { "type": "number" },
                "probability": { "type": "number" },
                "expectedClose": { "type": "string" },
                "createdAt": { "type": "string" },
                "updatedAt": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/email/send": {
      "post": {
        "summary": "Send email via Resend",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to", "subject", "body"],
                "properties": {
                  "to": { "type": "string", "format": "email" },
                  "subject": { "type": "string" },
                  "body": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/EmailSendResponse" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/email/draft": {
      "post": {
        "summary": "Generate AI email draft",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["transcript"],
                "properties": {
                  "transcript": { "type": "string" },
                  "leadContext": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "role": { "type": "string" },
                      "company": { "type": "string" }
                    }
                  },
                  "apiKey": { "type": "string" },
                  "model": { "type": "string", "enum": ["openai", "anthropic"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft generated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/EmailDraft" } }
            }
          },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api-keys": {
      "post": {
        "summary": "Create an API key",
        "description": "Creates a new read-only API key. The full key is returned only once.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Optional label for the key" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyNew" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Requires Pro plan" }
        }
      },
      "get": {
        "summary": "List API keys",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of API keys (full key never returned)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "keys": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ApiKeySummary" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api-keys/{keyHash}": {
      "delete": {
        "summary": "Revoke an API key",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "keyHash",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "revoked": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Key not found" }
        }
      }
    },
    "/sync": {
      "post": {
        "summary": "Sync derived data (summaries, action items, lead scores — never transcripts)",
        "description": "Opt-in: pushes derived meeting intelligence from the local-first client to the server so the read-only API can serve it. Body is idempotent; each item is upserted by id.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SyncPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synced",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Requires Pro plan" }
        }
      }
    },
...
  }
}